+++ src/main.zig
@@ -15,7 +15,7 @@ pub const writeZfi = index_format.writeZfi;
 pub const validateFasta = indexer.validateFasta;
 pub const scanHeaders = indexer.scanHeaders;
 
-const VERSION = "0.2.0";
+const VERSION = "0.2.3";
 
 const USAGE =
     \\usage: z-fasta <command> [options]